Release version 0.1.2

+++ eventsourcing-core/src/main/java/de/otto/edison/eventsourcing/kinesis/RetryPutRecordsKinesisClient.java
@@ -67,7 +67,7 @@ public class RetryPutRecordsKinesisClient {
 
     private void waitDependingOnRetryStep(int retryStep) {
         try {
-            Thread.sleep((int)Math.pow(2, retryStep) * 1000);
+            Thread.sleep((long)Math.pow(2, retryStep) * 1000);
         } catch (InterruptedException e) {
             e.printStackTrace();
         }
